Move XServer stuff to a ViewModel.
