+++ app/src/main/java/com/OxGames/Pluvia/PluviaApp.kt
@@ -4,7 +4,6 @@ import android.os.StrictMode
 import androidx.navigation.NavController
 import com.OxGames.Pluvia.events.EventDispatcher
 import com.google.android.play.core.splitcompat.SplitCompatApplication
-import com.winlator.xenvironment.XEnvironment
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
 
@@ -40,8 +39,5 @@ class PluviaApp : SplitCompatApplication() {
     companion object {
         internal val events: EventDispatcher = EventDispatcher()
         internal var onDestinationChangedListener: NavChangedListener? = null
-
-        // TODO: find a way to make this saveable, this is terrible
-        internal var xEnvironment: XEnvironment? = null
     }
 }
